Refactor the two timeout working tests to remove duplicate code
